# Witness-based pruning