@zsunberg 2024-07-29
Is it possible to keep multiple pruning strategies as options? (I have not looked at the code, so perhaps you have already implemented that?)