@qhho 2024-07-30
> Is it possible to keep multiple pruning strategies as options? (I have not looked at the code, so perhaps you have already implemented that?)

We can definitely do that. I briefly discussed with @WhiffleFish and we thought that having multiple pruning strategies isn't necessary if one performs better. I think with some code optimization, we can get witness-based pruning to generally perform better than the current pruning technique. That said, the original SARSOP paper does not mention witness-based pruning even though the code implementation uses it, so it might be a good idea to have both as options for completeness.