@dylan-asmar 2025-03-11
@qhho I haven't spent a lot of time looking at the code yet, but is this PR still compatible after PRs #18 and #19? Anything I can help out with?